refactured binary update from git - planing to use the same code in the async version as well.
